+++ bin/.gitignore
@@ -1,0 +1,2 @@
+/gitKaulins.class
+/GitKaulins.class
